Revert "cpufreq: Move get_cpu_idle_time() to cpufreq.c"
* Conflicts with our old dm_hotplug driver

This reverts commit c5d6114348a4d6d7e4ef7743bf7832fb8ce4b529.
